PR: Create File+Hashable, symlink Array+SwiftLint, fix nits
